improving the code
